% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/exploratory.R
\name{bc.ci.boot}
\alias{bc.ci.boot}
\title{Bootstrap confidence interval for sample statistics.}
\usage{
bc.ci.boot(data, symmetric = F, alpha = 0.05, B = 9999,
  show.progress = T)
}
\arguments{
\item{data}{Vector of angles.}

\item{symmetric}{Boolean indicating whether or not the data is assumed to be symmetric.}

\item{alpha}{Significance level of confidence interval to be obtained. Default is 0.05 (95% confidence interval).}

\item{B}{Number of bootstrap samples to use to estimate the confidence intervals. Default is 9999.}

\item{show.progress}{Boolean indicating whether or not to display a progress bar as the bootstrap is run.}
}
\value{
A list containing estimates of mu (location), rho (concentration), beta2 (skew) and alpha2 (kurtosis), and the upper and lower bounds of the confidence intervals calculated.
}
\description{
Obtain the main circular sample statistics of the distribution, with large-sample normal-theory confidence intervals.
}
\examples{
ests <- bc.ci.boot(q)
}

