% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exploratory.R
\name{bc.ci.LS}
\alias{bc.ci.LS}
\title{Large-sample confidence interval for sample statistics.}
\usage{
bc.ci.LS(data, alpha = 0.05, symmetric = F)
}
\arguments{
\item{data}{Vector of angles.}

\item{alpha}{Significance level of confidence interval to be obtained. Default is 0.05 (95\% confidence interval).}

\item{symmetric}{Boolean indicating whether or not the data is assumed to be symmetric.}
}
\value{
A list containing estimates of mu (location), rho (concentration), beta2 (skew) and alpha2 (kurtosis), and the upper and lower bounds of the confidence intervals calculated.
}
\description{
Obtain the main circular sample statistics of the distribution, with large-sample normal-theory confidence intervals.
}
\examples{
bar <- bc.ci.LS(q)
}
