% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/exploratory.R
\name{bc.sample.statistics}
\alias{bc.sample.statistics}
\title{Bias-corrected circular sample statistics.}
\usage{
bc.sample.statistics(data, symmetric = F)
}
\arguments{
\item{data}{Vector of angles.}

\item{symmetric}{Boolean indicating whether or not the data is assumed to be symmetric.}
}
\value{
A list containing estimates of mu (location), rho (concentration), beta2 (skew) and alpha2 (kurtosis).
}
\description{
Obtain the main circular sample statistics of the distribution, with large-sample normal-theory confidence intervals.
}
\examples{
bar <- bc.sample.statistics(q)
}

