% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exploratory.R
\name{circular.c.plot}
\alias{circular.c.plot}
\title{Plot of circular data}
\usage{
circular.c.plot(
  data,
  bins = 90,
  BW = 15,
  l.pos,
  l.size = 1,
  p.sep = 0.05,
  p.shrink = 1.5,
  yl = c(-1.2, 0.8),
  xl = c(-1.1, 0.9)
)
}
\arguments{
\item{data}{Vector of angles, in radians.}

\item{bins}{Specify number of bars to display in histogram. Default is 90.}

\item{BW}{Bandwidth to be used for the kernel density estiate. Default is 15.}

\item{l.pos}{Position of legend. Leave blank to plot without a legend.}

\item{l.size}{If legend is displayed, scale it by this factor. Default is 1.}

\item{p.sep}{Separation of points in plot. Default is 0.05}

\item{p.shrink}{Rescaling parameter of circular plot.}

\item{yl}{\code{ylim} parameter to be passed to circular plot.}

\item{xl}{\code{xlim} parameter to be passed to circular plot.}
}
\description{
Produces a circular plot of a directional data set, with kernel density (black), ML von Mises (blue) and ML Jones-Pewsey (red) distributions overlaid. The legend is optional.
}
\details{
See \code{\link{plot.circular}} for more on plotting parameters.
}
\examples{
circular.c.plot(q.4)
}
