% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Clustering.R
\name{cluster.pval.heatmap}
\alias{cluster.pval.heatmap}
\title{Plot cluster-by-cluster heatmap of p-values}
\usage{
cluster.pval.heatmap(m1, m2)
}
\arguments{
\item{m1}{Matrix of p-values to be plotted.}

\item{m2}{Optional: Second matrix of p-values. If supplied, the lower triangular matrix of \code{m1} is overwritten with that of \code{m2} and the two are plotted together.}
}
\description{
For a matrix of p-values, as generated by \code{\link{compare.cluster.dists}}, plot a heatmap highlighting p-values less than 0.05 or less than 0.1.
}
\examples{
cluster.pval.heatmap(res[,,1], res[,,2])
}

