% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Clustering.R
\name{cluster.sample.stats}
\alias{cluster.sample.stats}
\title{Sample statistics of subsets of data}
\usage{
cluster.sample.stats(data, clusts, include.noise = F)
}
\arguments{
\item{data}{Vector of angles.}

\item{clusts}{Vector of indices denoting the cluster to which each angle has been assigned.}

\item{include.noise}{Boolean indicator: should cluster 0 (noise cluster) be included in the output?}
}
\value{
A matrix containing the index and size of the subset tested; sample statistics as generated by \code{\list{bc.sample.statistics}}; von Mises and Jones-Pewsey maximum-likelihood estimators; AIC and BIC scores for ML von Mises and Jones-Pewsey distributions.
}
\description{
Divide an angular data set into subsets and obtain sample statistics for each subset.
}
\examples{
c.stats <- cluster.sample.stats(q, dbscan(centres, eps = 2.5, MinPts = 4)$cluster)
}

