% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Clustering.R
\name{compare.cluster.dists}
\alias{compare.cluster.dists}
\title{Compare all clusters for similar distribution}
\usage{
compare.cluster.dists(data, clusts, reps = 999, show.progress = F)
}
\arguments{
\item{data}{Vector of raw angles (not quartered/wrapped angles).}

\item{clusts}{Vector of indices denoting the cluster to which each angle has been assigned.}

\item{reps}{Number of randomisation tests to apply}
}
\value{
A three-dimensional array containing p-values for tests of similar mean, concentration and distribution.
}
\description{
Divide an angular data set into subsets, then subdivide those into perpendicular quadrants and obtain sample statistics for each subset.
}
\examples{
db <- dbscan(centres, eps = 2.5, MinPts = 4)$cluster
res <- compare.cluster.dists(q.4, db, reps = 99)
}

