% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Clustering.R
\name{compare.clusters}
\alias{compare.clusters}
\title{Compare subsets of data against others}
\usage{
compare.clusters(data, clusts, include.noise = F, reps = 999)
}
\arguments{
\item{data}{Vector of angles.}

\item{clusts}{Vector of indices denoting the cluster to which each angle has been assigned.}

\item{include.noise}{Boolean indicator: should cluster 0 (noise cluster) be included in the output?}
}
\value{
A matrix containing the size of the subset tested; the Rayleigh test score (\code{\link{rayleigh.test}}); Watson mean test score (\code{\link{watson.common.mean.test}}); Wallraff concentration test score (\code{\link{wallraff.concentration.test}});
Mardia-Watson-Wheeler large-sample test of common distribution (\code{\link{mww.common.dist.LS}}); and randomised Watson two-sample test of common distribution (\code{\link{same.dist.watson.rand}}).
}
\description{
Divide an angular data set into subsets and compare the characteristics of each subset against the remaining angles, to test whether they might plausibly share the same distribution.
}
\examples{
c.res <- compare.clusters(q, dbscan(centres, eps = 2.5, MinPts = 4)$cluster)
}

