% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two-sample-tests.R
\name{cs.unif.scores}
\alias{cs.unif.scores}
\title{Cosine & sine rank scores}
\usage{
cs.unif.scores(samples)
}
\arguments{
\item{samples}{A list containing the vectors of angles (in radians) to be evaluated.}
}
\value{
List containing the cosine and sine rank scores.
}
\description{
Support function: calculate cosine & sine rank scores for a sample of angles
}
\examples{
cs.scores <- cs.unif.scores(list(q.4.a, q.4.b))
}
