% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Clustering.R
\name{EM.clusters}
\alias{EM.clusters}
\title{Classify points using EM mixture model}
\usage{
EM.clusters(data, model)
}
\arguments{
\item{data}{Set of angular data}

\item{model}{Fitted mixture model from \code{\link{EM.vonmises}}.}
}
\value{
Vector of cluster numbers
}
\description{
Winner-takes-all classification of points based on a von Mises mixture model obtained using \code{\link{EM.vonmises}}.
}
\examples{
ex1 <- c(rvonmises(120 * 0.3, mu = circular(pi/2), kappa = 10),
         rvonmises(120 * 0.7, mu = circular(pi), kappa = 3))
em1 <- EM.vonmises(ex1, k = 2)
cl <- EM.clusters(ex1, em1)
plot(ex1[cl == 1], stack = T)
points(ex1[cl == 2], col = "blue", stack = T)
}

