% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exploratory.R
\name{get.moments}
\alias{get.moments}
\title{Circular sample moments}
\usage{
get.moments(data)
}
\arguments{
\item{data}{Vector of angles.}
}
\value{
A list containing the sample mean, mean resultant length, and second, third and fourth sine and consine moments of the data.
}
\description{
Obtain the key sample trigonometric moments of the distribution.
}
\examples{
bar <- get.moments(q)
}
