% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jones-Pewsey.R
\name{JP.ci.nt}
\alias{JP.ci.nt}
\title{Normal-theory confidence interval for Jones-Pewsey parameter estimators}
\usage{
JP.ci.nt(jp.ests, alpha = 0.05)
}
\arguments{
\item{jp.ests}{List (as output by \code{JP.mle}) containing named estimates of \code{mu}, \code{kappa} and \code{psi}, and the Hessian matrix \code{HessMat} of the optimised parameters.}

\item{alpha}{Significance level of confidence interval to be obtained. Default is 0.05 (95\% confidence interval).}
}
\value{
A list containing estimates of \code{mu}, \code{kappa} and \code{psi}, and the upper and lower bounds of the confidence intervals calculated.
}
\description{
Obtain nominal 100(1-\code{alpha})% confidence intervals for Jones-Pewsey estimators, by solving the Hessian matrix produced by \code{JP.mle}.
}
\examples{
ci <- JP.ci.nt(JP.mle(k.2), alpha = 0.05)
}
