% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Jones-Pewsey.R
\name{JP.df}
\alias{JP.df}
\title{Jones-Pewsey distribution function}
\usage{
JP.df(theta, mu, kappa, psi, ncon)
}
\arguments{
\item{theta}{Angle, in radians, at which the distribution is to be calculated. Can be either numeric or vector.}

\item{mu}{Mean direction parameter.}

\item{kappa}{Concentration parameter.}

\item{psi}{Shape parameter.}
}
\value{
The distribution function evaluated at \code{theta}: either a numeric or a vector, depending on the input.
}
\description{
Calculate the cdf of the Jones-Pewsey distribution at given points, based on specified values of \code{mu}, \code{kappa} and \code{psi}.
}
\examples{
cdf <- JP.df(theta = c(k.2), mu = 3, kappa = 1, psi = -3, JP.NCon(kappa = 1, psi = -3))
}

