% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jones-Pewsey.R
\name{JP.GoF}
\alias{JP.GoF}
\title{Goodness-of-fit tests for Jones-Pewsey distribution}
\usage{
JP.GoF(data, mu, kappa, psi, display = T)
}
\arguments{
\item{data}{Vector of angles to be tested against the Jones-Pewsey distribution.}

\item{mu}{Mean direction parameter of the Jones-Pewsey distribution to be fitted.}

\item{kappa}{Concentration parameter of the Jones-Pewsey distribution to be fitted.}

\item{psi}{Shape parameter of the Jones-Pewsey distribution to be fitted.}

\item{display}{Boolean specifying whether to print the results of the four tests to the console or not.}
}
\value{
List of test statistics for the four tests, plus the p-value for the Rayleigh tests.
}
\description{
Runs a bundle of goodness-of-fit tests on the data against the Jones-Pewsey distribution.
}
\details{
The tests rely on doubling the Jones-Pewsey distribution function of each angle in the data set, and testing the resulting distribution for uniformity.
}
\examples{
GoF <- JP.GoF(q, mu, kappa, psi)
}
\seealso{
The individual tests included in the bundle: \code{\link{kuiper.test}}, \code{\link{watson.test}}, \code{\link{rao.spacing.test}}, and \code{\link{rayleigh.test}}.
}
