% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jones-Pewsey.R
\name{JP.mle}
\alias{JP.mle}
\title{Maximum likelihood estimator for Jones-Pewsey parameters}
\usage{
JP.mle(data)
}
\arguments{
\item{data}{Vector of angles over which maximum likelihood estimation is to be performed.}
}
\value{
A list containing the maximum likelihood achieved, estimates of \code{mu}, \code{kappa} and \code{psi}, and the Hessian matrix of the optimised parameters.
}
\description{
Obtain maximum likelihood estimates of Jones-Pewsey parameters for a given data set.
}
\examples{
mu <- JP.mle(k.2)$mu
}
