% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Jones-Pewsey.R
\name{JP.NCon}
\alias{JP.NCon}
\title{Jones-Pewsey normalising constant}
\usage{
JP.NCon(kappa, psi)
}
\arguments{
\item{kappa}{Concentration parameter.}

\item{psi}{Shape parameter.}
}
\value{
A constant value to be passed to further functions.
}
\description{
Support function to calculate the normalising constant for the Jones-Pewsey distribution, based on specified values of \code{kappa} and \code{psi}.
}
\examples{
nc <- JP.NCon(kappa = 1, psi = -3)
}

