% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jones-Pewsey.R
\name{JP.PP}
\alias{JP.PP}
\title{Jones-Pewsey P-P plot}
\usage{
JP.PP(data, mu, kappa, psi)
}
\arguments{
\item{data}{Vector of angles to be fitted against the Jones-Pewsey distribution.}

\item{mu}{Mean direction parameter.}

\item{kappa}{Concentration parameter.}

\item{psi}{Shape parameter.}
}
\description{
Produces a P-P plot of the data against a specified Jones-Pewsey distribution, to graphically assess the goodness of fit of the model.
}
\examples{
JP.PP(k.2, mu = 3, kappa = 1, psi = -3)
}
