% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Jones-Pewsey.R
\name{JP.psi.info}
\alias{JP.psi.info}
\title{AIC and BIC for Jones-Pewsey against nested models}
\usage{
JP.psi.info(data, psi.0 = 0)
}
\arguments{
\item{data}{Vector of angles to be tested.}

\item{psi.0}{Value of \code{psi} specifying a simpler model, to test against the three-parameter Jones-Pewsey.}
}
\value{
List containing a matrix ofthe parameters tested and their associated AIC and BIC scores, as well as a comment explaining the output.
}
\description{
Obtain the AIC and BIC scores for Jones-Pewsey distribution with MLE parameters over a simpler model with a specified value of \code{psi}.
}
\details{
von Mises has psi -> 0; cardioid has psi = 1; wrapped Cauchy has psi = -1.
}
\examples{
JP.info <- JP.psi.info(q)
}
\seealso{
\code{\link{AIC}}, \code{\link{BIC}}
}

