% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Jones-Pewsey.R
\name{JP.psi.LR.boot}
\alias{JP.psi.LR.boot}
\title{Bootstrap likelihood ratio test for Jones-Pewsey against nested models}
\usage{
JP.psi.LR.boot(data, psi.0 = 0, B = 9999, alpha = 0.05,
  show.progress = T)
}
\arguments{
\item{data}{Vector of angles to be tested.}

\item{psi.0}{Value of \code{psi} specifying a simpler model, to test against the three-parameter Jones-Pewsey.}

\item{B}{Number of bootstrap samples to use to obtain the confidence interval. Default is 9999.}

\item{alpha}{Significance level of likelihood ratio test. Default is 0.05 (95% confidence interval).}
}
\value{
List containing the deviance, p-value, matrix ofthe parameters tested, and a comment explaining the output.
}
\description{
Bootstrap likelihood ratio test of Jones-Pewsey distribution with MLE parameters against a simpler model with a specified value of \code{psi}.
}
\details{
von Mises has psi -> 0; cardioid has psi = 1; wrapped Cauchy has psi = -1.
}
\examples{
JP.LR.test <- JP.psi.LR.boot(q, psi.0 = 1, B = 99)
}

