% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jones-Pewsey.R
\name{JP.psi.LR.test}
\alias{JP.psi.LR.test}
\title{Likelihood ratio test for Jones-Pewsey against nested models}
\usage{
JP.psi.LR.test(data, psi.0 = 0, alpha = 0.05)
}
\arguments{
\item{data}{Vector of angles to be tested.}

\item{psi.0}{Value of \code{psi} specifying a simpler model, to test against the three-parameter Jones-Pewsey.}

\item{alpha}{Significance level of likelihood ratio test. Default is 0.05 (95\% confidence interval).}
}
\value{
List containing the deviance, p-value, matrix ofthe parameters tested, and a comment explaining the output.
}
\description{
Likelihood ratio test of Jones-Pewsey distribution with MLE parameters against a simpler model with a specified value of \code{psi}.
}
\details{
von Mises has psi -> 0; cardioid has psi = 1; wrapped Cauchy has psi = -1.
}
\examples{
JP.LR.test <- JP.psi.LR.test(q, psi.0 = 1)
}
