% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jones-Pewsey.R
\name{JP.qf}
\alias{JP.qf}
\title{Jones-Pewsey quantile function}
\usage{
JP.qf(u, mu, kappa, psi, ncon)
}
\arguments{
\item{u}{Quantile to be evaluated.}

\item{mu}{Mean direction parameter.}

\item{kappa}{Concentration parameter.}

\item{psi}{Shape parameter.}
}
\value{
The quantile function evaluated at \code{u}: either a numeric or a vector, depending on the input.
}
\description{
Evaluate the quartile of the Jones-Pewsey distribution at given points, based on specified values of \code{mu}, \code{kappa} and \code{psi}.
}
\examples{
q <- JP.qf(u = 0.9, mu = 3, kappa = 1, psi = -3, JP.NCon(kappa = 1, psi = -3))
}
