% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Jones-Pewsey.R
\name{JP.QQ}
\alias{JP.QQ}
\title{Jones-Pewsey Q-Q plot of data}
\usage{
JP.QQ(data, mu, kappa, psi)
}
\arguments{
\item{data}{Vector of angles to be fitted against the Jones-Pewsey distribution.}

\item{mu}{Mean direction parameter.}

\item{kappa}{Concentration parameter.}

\item{psi}{Shape parameter.}
}
\description{
Produces a Q-Q plot of the data against a specified Jones-Pewsey distribution, to graphically assess the goodness of fit of the model.
}
\examples{
JP.QQ(k.2, mu = 3, kappa = 1, psi = -3)
}

