% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Jones-Pewsey.R
\name{JP.sim}
\alias{JP.sim}
\title{Jones-Pewsey simulation function}
\usage{
JP.sim(n, mu, kappa, psi, ncon)
}
\arguments{
\item{n}{Number of angles to be simulated from the distribution.}

\item{mu}{Mean direction parameter.}

\item{kappa}{Concentration parameter.}

\item{psi}{Shape parameter.}
}
\value{
A vector of n simulated angles from the specified distribution.
}
\description{
Simulate a number of angles from the Jones-Pewsey distribution specified by \code{mu}, \code{kappa} and \code{psi}.
}
\examples{
sample <- JP.sim(n = 100, mu = 3, kappa = 1, psi = -3, JP.NCon(kappa = 1, psi = -3))
}

