% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Clustering.R
\name{MSclust.p.est}
\alias{MSclust.p.est}
\title{Estimate stabilising parameter p for mean-shift clustering of circular data}
\usage{
MSclust.p.est(data, dp = 4, plot = T)
}
\arguments{
\item{data}{Set of angular data to be clustered.}

\item{dp}{How many decimal places of convergence is required before p is accepted?}

\item{plot}{Boolean: plot the correlation curve or not? Default is T.}
}
\value{
Suggested value of p to be used in mean-shift clustering functions such as \code{\link{MSclust.NB}}
}
\description{
Automation of graphical method of estimating stabilising parameter to be used in mean-shift clustering algorithms. Calculates correlation of kernel density estimate for successive values of p, selecting as the optimum the maximum value of p that is less than 1 when rounded to a certain degree of precision.
}
\examples{
ex1 <- c(rvonmises(120 * 0.3, mu = circular(pi/2), kappa = 10),
         rvonmises(120 * 0.7, mu = circular(pi), kappa = 3))
p <- MSclust.p.est(ex1, dp = 4)
}

