% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Two-sample-tests.R
\name{mww.common.dist.LS}
\alias{mww.common.dist.LS}
\title{Test for common distribution among multiple samples}
\usage{
mww.common.dist.LS(cs.scores, sample.sizes)
}
\arguments{
\item{cs.scores}{A list containing sine and cosine rank scores, output from the data of interest by \code{\link{cs.unif.scores}}.}

\item{sample.sizes}{A vector defining the sizes of the samples that were used to produce the \code{cs.scores} object.}
}
\value{
List containing the test statistic and p-value.
}
\description{
Large-sample Mardia-Wheeler-Watson test for common distribution.
}
\examples{
cd.res <- mww.common.dist.LS(cs.unif.scores(list(q.4.a, q.4.b)), c(length(q.4.a), length(q.4.b)))
}

