% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two-sample-tests.R
\name{mww.common.dist.rand}
\alias{mww.common.dist.rand}
\title{Permutation test for common distribution among multiple samples}
\usage{
mww.common.dist.rand(cs.scores, sample.sizes, NR = 9999, show.progress = T)
}
\arguments{
\item{cs.scores}{A list containing sine and cosine rank scores, output from the data of interest by \code{\link{cs.unif.scores}}.}

\item{sample.sizes}{A vector defining the sizes of the samples that were used to produce the \code{cs.scores} object.}

\item{NR}{Number of permutation samples to use to estimate the p-value. Default is 9999.}

\item{show.progress}{Boolean indicating whether or not to display a progress bar as the bootstrap is run.}
}
\value{
p-value for the test.
}
\description{
Permutation-based Mardia-Wheeler-Watson test for common distribution, to be used when any of the samples has less than 10 angles.
}
\examples{
cd.res <- mww.common.dist.rand(list(q.4.a, q.4.b))
}
