% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Clustering.R
\name{plot.EM.vonmises}
\alias{plot.EM.vonmises}
\title{Plot mixture von Mises model found by EM algorithm.}
\usage{
plot.EM.vonmises(varToPlot, modelToPlot)
}
\arguments{
\item{varToPlot}{Set of angular data}

\item{modelToPlot}{Fitted mixture model from \code{\link{EM.vonmises}}.}
}
\description{
Plot original data with individual components and fitted mixture model.
}
\examples{
ex1 <- c(rvonmises(120 * 0.3, mu = circular(pi/2), kappa = 10),
         rvonmises(120 * 0.7, mu = circular(pi), kappa = 3))
em1 <- EM.vonmises(ex1, k = 2)
plot.EM.vonmises(ex1, em1)
}

