% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Two-sample-tests.R
\name{pooled.mean}
\alias{pooled.mean}
\title{Calculate a pooled mean for two or more samples}
\usage{
pooled.mean(samples, alpha = 0.05)
}
\arguments{
\item{samples}{A list containing the vectors of angles (in radians) to be tested.}

\item{alpha}{Significance level of confidence interval to be obtained. Default is 0.05 (95pc)}
}
\value{
List containing the observed value of the test statistic and the p-value.
}
\description{
Based on Fisher's pooled estimate of a common mean for two or more samples of angular data.
}
\examples{
m <- pooled.mean((list(q.4.a, q.4.b))
}

