% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/exploratory.R
\name{r.symm.test.boot}
\alias{r.symm.test.boot}
\title{Bootstrap reflective symmetry test}
\usage{
r.symm.test.boot(data, B = 9999, alpha = 0.05, show.progress = T)
}
\arguments{
\item{data}{Vector of angles.}

\item{B}{Number of bootstrap samples to use to estimate the standard error of the p-value. Default is 9999.}

\item{alpha}{Significance level of confidence interval to be obtained. Default is 0.05 (95pc confidence interval).}

\item{show.progress}{Boolean indicating whether or not to display a progress bar as the bootstrap is run.}
}
\value{
A list containing the p-value of the test and its standard error.
}
\description{
Bootstrap test of reflective symmetry about an unspecified mean.
}
\examples{
reject.symmetry <- r.symm.test.boot(q)
}
\seealso{
\code{\link{wilcox.test}} for a rank test of reflective symmetry about a specified direction.
}

