% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exploratory.R
\name{r.symm.test.stat}
\alias{r.symm.test.stat}
\title{Reflective symmetry test}
\usage{
r.symm.test.stat(data)
}
\arguments{
\item{data}{Vector of angles.}

\item{alpha}{Significance level of confidence interval to be obtained. Default is 0.05 (95\% confidence interval).}

\item{symmetric}{Boolean indicating whether or not the data is assumed to be symmetric.}
}
\value{
A list containing the test statistic and p-value of the test.
}
\description{
Large-sample asymptotic-theory test of reflective symmetry about an unspecified mean.
}
\examples{
reject.symmetry <- r.symm.test.stat(q)$p.value < 0.05
}
\seealso{
\code{\link{wilcox.test}} for a rank test of reflective symmetry about a specified direction.
}
