% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Two-sample-tests.R
\name{two.sample.QQ}
\alias{two.sample.QQ}
\title{Two-sample Q-Q plot}
\usage{
two.sample.QQ(data1, data2)
}
\arguments{
\item{data1}{First vector of angles to be tested.}

\item{data2}{Second vector of angles to be tested.}
}
\description{
Plot the quartiles of two distributions against each other: graphical test of similarity of fit
}
\examples{
two.sample.QQ(q.a, q.b)
}

