% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exploratory.R
\name{uniformity.plot}
\alias{uniformity.plot}
\title{Uniformity plot of data}
\usage{
uniformity.plot(data, extend = F)
}
\arguments{
\item{data}{Vector of angles, in radians.}

\item{extend}{Boolean specifying whether to extend the tails of the plot to make any pattern at the extremes of the data easier to see.}
}
\description{
Produces a uniform P-P plot. Graphical tool to assess uniformity of data.
}
\examples{
uniformity.plot(q)
}
