% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exploratory.R
\name{uniformity.tests}
\alias{uniformity.tests}
\title{Bundle of tests of uniformity}
\usage{
uniformity.tests(data, display = T)
}
\arguments{
\item{data}{Vector of angles.}

\item{display}{Boolean specifying whether to print the results of the four tests to the console or not.}
}
\value{
List of test statistics for the four tests, plus the p-value for the Rayleigh tests.
}
\description{
Runs a bundle of goodness-of-fit tests on the data against the uniform distribution.
}
\examples{
bar$mu <- bc.point.estimates(q)$mu
}
\seealso{
The individual tests included in the bundle: \code{\link{kuiper.test}}, \code{\link{watson.test}}, \code{\link{rao.spacing.test}}, and \code{\link{rayleigh.test}}.
}
