% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/vonMises.R
\name{vM.GoF.boot}
\alias{vM.GoF.boot}
\title{Bootstrap goodness-of-fit tests for von Mises distribution}
\usage{
vM.GoF.boot(data, B = 9999, show.progress = T)
}
\arguments{
\item{data}{Vector of angles to be tested against the von Mises distribution.}

\item{B}{Number of bootstrap samples to use to obtain the confidence interval. Default is 9999.}

\item{show.progress}{Boolean indicating whether or not to display a progress bar as the bootstrap is run.}
}
\value{
List of p-values for the four tests.
}
\description{
Runs a bundle of bbootstrap goodness-of-fit tests on the data against the von Mises distribution.
}
\details{
The tests rely on doubling the von Mises distribution function of each angle in the data set, and testing the resulting distribution for uniformity.
}
\examples{
GoF.boot <- vM.GoF.boot(q, B = 99)
}
\seealso{
The individual tests included in the bundle: \code{\link{kuiper.test}}, \code{\link{watson.test}}, \code{\link{rao.spacing.test}}, and \code{\link{rayleigh.test}}.
}

