% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonMises.R
\name{vM.GoF}
\alias{vM.GoF}
\title{Goodness-of-fit tests for von Mises distribution}
\usage{
vM.GoF(data, mu, kappa, display = T)
}
\arguments{
\item{data}{Vector of angles to be tested against the von Mises distribution.}

\item{mu}{Mean direction parameter of the von Mises distribution to be fitted.}

\item{kappa}{Concentration parameter of the von Mises distribution to be fitted. Must be between 0 and 1.}

\item{display}{Boolean specifying whether to print the results of the four tests to the console or not.}
}
\value{
List of test statistics for the four tests, plus the p-value for the Rayleigh tests.
}
\description{
Runs a bundle of goodness-of-fit tests on the data against a specified von Mises distribution.
}
\details{
Four of the tests rely on doubling the von Mises distribution function of each angle in the data set, and testing the resulting distribution for uniformity.
}
\examples{
GoF <- vM.GoF(q, mu, kappa)
}
\seealso{
The individual tests included in the bundle: \code{\link{kuiper.test}}, \code{\link{watson.test}}, \code{\link{rao.spacing.test}}, and \code{\link{rayleigh.test}}.
}
