% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two-sample-tests.R
\name{wallraff.concentration.test}
\alias{wallraff.concentration.test}
\title{Wallraff's two-sample test of common concentration}
\usage{
wallraff.concentration.test(samples)
}
\arguments{
\item{samples}{A list containing the vectors of angles (in radians) to be tested.}
}
\value{
List containing the test statistic, p-value and ratio of maximum to minimum sample dispersion.
}
\description{
Wallraff's nonparametric test of circular homoscedasticity (nonparametric test for similar concentration of multiple samples).
}
\examples{
comp <- watson.common.mean.test(list(q.4.a, q.4.b))
}
