% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Two-sample-tests.R
\name{watson.common.mean.test}
\alias{watson.common.mean.test}
\title{Watson's two-sample test of common mean}
\usage{
watson.common.mean.test(samples)
}
\arguments{
\item{samples}{A list containing the vectors of angles (in radians) to be tested.}
}
\value{
List containing the test statistic, p-value and ratio of maximum to minimum sample dispersion.
}
\description{
Watson's large-sample nonparametric test for the similarity of a common mean for two (or more) distributions. Does not assume that the distributions tested have a common concentration or shape.
}
\examples{
comp <- watson.common.mean.test(list(q.4.a, q.4.b))
}

