% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two-sample-tests.R
\name{watson.mean.test.boot}
\alias{watson.mean.test.boot}
\title{Bootstrap version of Watson's two-sample test of common mean}
\usage{
watson.mean.test.boot(samples, B = 9999, show.progress = T)
}
\arguments{
\item{samples}{A list containing the vectors of angles (in radians) to be tested.}

\item{B}{Number of bootstrap samples to use to obtain the estimated p-value. Default is 9999.}

\item{show.progress}{Boolean indicating whether or not to display a progress bar as the bootstrap is run.}
}
\value{
List containing the test statistic, p-value and ratio of maximum to minimum sample dispersion.
}
\description{
Bootstrap of Watson's nonparametric test for the similarity of a common mean for two (or more) distributions. Does not assume that the distributions tested have a common concentration or shape.
}
\examples{
comp <- watson.mean.test.boot(list(q.4.a, q.4.b), B = 99)
}
