% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Two-sample-tests.R
\name{watson.two.test.rand}
\alias{watson.two.test.rand}
\title{Permutation test for common distribution between two samples}
\usage{
watson.two.test.rand(data1, data2, NR = 9999, show.progress = T)
}
\arguments{
\item{data1}{First vector of angles to be tested.}

\item{data2}{Second vector of angles to be tested.}

\item{NR}{Number of permutation samples to use to estimate the p-value. Default is 9999.}

\item{show.progress}{Boolean indicating whether or not to display a progress bar as the bootstrap is run.}
}
\value{
List containing the observed value of the test statistic and the p-value.
}
\description{
Extension of \code{\link{watson.two.test}}, using permutation to obtain the p-value.
}
\examples{
cd.res <- watson.two.test.rand(list(q.4.a, q.4.b))
}
\seealso{
\code{\link{watson.two.test}}
}

